Create index selection bar
Much like the parameter selection bars, the index selector presents
a dropdown list of index files that can be used. Files are also uploaded
through the index selector.

A new, higher-level component has been created to store the mugs reader.
It currently contains the index selector and five standard selector bars,
one for each grade (and miscellanous grades). This component has been
assigned to the root for the time being.